+++ DfE.FindInformationAcademiesTrusts/Pages/ViewConstants.cs
@@ -28,27 +28,27 @@ public static class ViewConstants
 
     public static readonly List<ExternalServiceLink> ExternalServiceLinks =
     [
-        new ExternalServiceLink("Prepare conversions and transfers",
+        new("Prepare conversions and transfers",
             "Create a transfer or conversion project document for an advisory board.",
             "https://educationgovuk.sharepoint.com/sites/lvewp00299/SitePages/Prepare-Conversions-and-Transfers.aspx"),
 
-        new ExternalServiceLink("Complete conversions, transfers and changes",
+        new("Complete conversions, transfers and changes",
             "Manage a conversion or transfer project after it has been to an advisory board.",
             "https://educationgovuk.sharepoint.com/sites/lvewp00299/SitePages/complete-conversions-transfers-and-changes.aspx"),
 
-        new ExternalServiceLink("Record concerns or support for trusts",
+        new("Record concerns or support for trusts",
             "Add cases or interactions, record risks and log support and concerns for trusts.",
             "https://educationgovuk.sharepoint.com/sites/lvewp00299/SitePages/Record-concerns-and-support-for-trusts.aspx"),
 
-        new ExternalServiceLink("Manage free school projects",
+        new("Manage free school projects",
             "Manage presumption projects in the pre-opening phase.",
             "https://educationgovuk.sharepoint.com/sites/lvewp00299/SitePages/Manage-free-school-projects.aspx?web=1"),
 
-        new ExternalServiceLink("Reporting and data tools",
+        new("Reporting and data tools",
             "Tools to help you gather educational data and create reports.",
             "https://educationgovuk.sharepoint.com/sites/lvewp00299/SitePages/reportinganddata.aspx"),
 
-        new ExternalServiceLink("High quality trust framework",
+        new("High quality trust framework",
             "Process guidance and tools for making trust-related project decisions.",
             "https://educationgovuk.sharepoint.com/sites/lvewp00299/SitePages/RG%20high%20quality%20trust%20framework.aspx")
     ];
@@ -63,9 +63,9 @@ public static class ViewConstants
     public const string ContactsInTrustPageName = "In the trust";
 
     public const string OfstedPageName = "Ofsted";
+    public const string OfstedSingleHeadlineGradesPageName = "Single headline grades";
     public const string OfstedCurrentRatingsPageName = "Current ratings";
     public const string OfstedPreviousRatingsPageName = "Previous ratings";
-    public const string OfstedImportantDatesPageName = "Important dates";
     public const string OfstedSafeguardingAndConcernsPageName = "Safeguarding and concerns";
 
     public const string GovernancePageName = "Governance";